+++ backend/src/main/java/zipgo/auth/application/AuthService.java
@@ -3,7 +3,7 @@ package zipgo.auth.application;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import zipgo.auth.OAuthResponse;
+import zipgo.auth.application.dto.OAuthResponse;
 import zipgo.auth.util.JwtProvider;
 import zipgo.member.domain.Member;
 import zipgo.member.domain.repository.MemberRepository;
