+++ backend/src/main/java/zipgo/auth/application/dto/KakaoInfoResponse.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.databind.annotation.JsonNaming;
 
 
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public record KakaoDetailDto(
+public record KakaoInfoResponse(
         String accessToken,
         String tokenType,
         String refreshToken,
