+++ backend/src/main/java/zipgo/auth/application/dto/KakaoOAuthResponse.java
@@ -1,4 +1,4 @@
-package zipgo.auth;
+package zipgo.auth.application.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
