+++ backend/src/test/java/zipgo/auth/application/AuthServiceTest.java
@@ -7,8 +7,8 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import zipgo.auth.KakaoOAuthResponse;
-import zipgo.auth.OAuthResponse;
+import zipgo.auth.application.dto.KakaoOAuthResponse;
+import zipgo.auth.application.dto.OAuthResponse;
 import zipgo.auth.util.JwtProvider;
 import zipgo.member.domain.Member;
 import zipgo.member.application.MemberQueryService;
